@@ -6,6 +6,7 @@
   "scripts": {
     "build": "tsc",
     "test": "jest",
+    "prepare": "tsc",
     "prepublishOnly": "npm run test && npm run build",
     "version": "standard-version"
   },
@@ -36,16 +37,14 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "^22.0.0",
-    "jest": "^22.0.4",
-    "standard-version": "^4.2.0",
-    "typescript": "^2.6.2"
+    "@types/jest": "^24.0.12",
+    "@types/node": "^12.0.0",
+    "jest": "^24.6.0",
+    "standard-version": "^6.0.1",
+    "typescript": "^3.4.5"
   },
   "dependencies": {
-    "extract-text-webpack-plugin": "^1.0.1",
-    "postcss-loader": "^2.0.9",
-    "style-loader": "^0.19.1",
     "typings-for-css-modules-loader": "^1.7.0",
-    "webpack": "^1.15.0"
+    "webpack": "^4.28.4"
   }
 }
