@@ -1,4 +1,4 @@
-exports.defaults = {
+export default {
   namedExport: true,
   modules: true,
   localIdentName: "[name]__[local]___[hash:base64:5]",
